ai-service-embedding/requirements.txt
flask
flask-cors
sentence-transformers
faiss-cpu
numpy
mysql-connector-python
